canbus: add aeb control command in devkit     Canbus: update steering command and report in devkit     Calibration:update kinetic calibration parameter for dev_kit_standard (#13141)

+++ modules/canbus/vehicle/devkit/protocol/steering_command_102.h
@@ -36,56 +36,56 @@ class Steeringcommand102 : public ::apollo::drivers::canbus::ProtocolData<
 
   void Reset() override;
 
-  // config detail: {'name': 'Steer_EN_CTRL', 'enum': {0:
-  // 'STEER_EN_CTRL_DISABLE', 1: 'STEER_EN_CTRL_ENABLE'}, 'precision': 1.0,
-  // 'len': 1, 'is_signed_var': False, 'offset': 0.0, 'physical_range': '[0|1]',
-  // 'bit': 0, 'type': 'enum', 'order': 'motorola', 'physical_unit': ''}
+  // config detail: {'bit': 0, 'enum': {0: 'STEER_EN_CTRL_DISABLE', 1:
+  // 'STEER_EN_CTRL_ENABLE'}, 'is_signed_var': False, 'len': 1, 'name':
+  // 'Steer_EN_CTRL', 'offset': 0.0, 'order': 'motorola', 'physical_range':
+  // '[0|1]', 'physical_unit': '', 'precision': 1.0, 'type': 'enum'}
   Steeringcommand102* set_steer_en_ctrl(
       Steering_command_102::Steer_en_ctrlType steer_en_ctrl);
 
-  // config detail: {'name': 'Steer_ANGLE_Target', 'offset': -500.0,
-  // 'precision': 0.1, 'len': 16, 'is_signed_var': False, 'physical_range':
-  // '[-500|500]', 'bit': 31, 'type': 'double', 'order': 'motorola',
-  // 'physical_unit': 'deg'}
-  Steeringcommand102* set_steer_angle_target(double steer_angle_target);
+  // config detail: {'bit': 31, 'is_signed_var': False, 'len': 16, 'name':
+  // 'Steer_ANGLE_Target', 'offset': -500.0, 'order': 'motorola',
+  // 'physical_range': '[-500|500]', 'physical_unit': 'deg', 'precision': 1.0,
+  // 'type': 'int'}
+  Steeringcommand102* set_steer_angle_target(int steer_angle_target);
 
-  // config detail: {'name': 'Steer_ANGLE_SPD', 'offset': 0.0, 'precision': 1.0,
-  // 'len': 8, 'is_signed_var': False, 'physical_range': '[0|250]', 'bit': 15,
-  // 'type': 'int', 'order': 'motorola', 'physical_unit': 'deg/s'}
+  // config detail: {'bit': 15, 'is_signed_var': False, 'len': 8, 'name':
+  // 'Steer_ANGLE_SPD', 'offset': 0.0, 'order': 'motorola', 'physical_range':
+  // '[0|250]', 'physical_unit': 'deg/s', 'precision': 1.0, 'type': 'int'}
   Steeringcommand102* set_steer_angle_spd(int steer_angle_spd);
 
-  // config detail: {'name': 'CheckSum_102', 'offset': 0.0, 'precision': 1.0,
-  // 'len': 8, 'is_signed_var': False, 'physical_range': '[0|255]', 'bit': 63,
-  // 'type': 'int', 'order': 'motorola', 'physical_unit': ''}
+  // config detail: {'bit': 63, 'is_signed_var': False, 'len': 8, 'name':
+  // 'CheckSum_102', 'offset': 0.0, 'order': 'motorola', 'physical_range':
+  // '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'int'}
   Steeringcommand102* set_checksum_102(int checksum_102);
 
  private:
-  // config detail: {'name': 'Steer_EN_CTRL', 'enum': {0:
-  // 'STEER_EN_CTRL_DISABLE', 1: 'STEER_EN_CTRL_ENABLE'}, 'precision': 1.0,
-  // 'len': 1, 'is_signed_var': False, 'offset': 0.0, 'physical_range': '[0|1]',
-  // 'bit': 0, 'type': 'enum', 'order': 'motorola', 'physical_unit': ''}
+  // config detail: {'bit': 0, 'enum': {0: 'STEER_EN_CTRL_DISABLE', 1:
+  // 'STEER_EN_CTRL_ENABLE'}, 'is_signed_var': False, 'len': 1, 'name':
+  // 'Steer_EN_CTRL', 'offset': 0.0, 'order': 'motorola', 'physical_range':
+  // '[0|1]', 'physical_unit': '', 'precision': 1.0, 'type': 'enum'}
   void set_p_steer_en_ctrl(
       uint8_t* data, Steering_command_102::Steer_en_ctrlType steer_en_ctrl);
 
-  // config detail: {'name': 'Steer_ANGLE_Target', 'offset': -500.0,
-  // 'precision': 0.1, 'len': 16, 'is_signed_var': False, 'physical_range':
-  // '[-500|500]', 'bit': 31, 'type': 'double', 'order': 'motorola',
-  // 'physical_unit': 'deg'}
-  void set_p_steer_angle_target(uint8_t* data, double steer_angle_target);
+  // config detail: {'bit': 31, 'is_signed_var': False, 'len': 16, 'name':
+  // 'Steer_ANGLE_Target', 'offset': -500.0, 'order': 'motorola',
+  // 'physical_range': '[-500|500]', 'physical_unit': 'deg', 'precision': 1.0,
+  // 'type': 'int'}
+  void set_p_steer_angle_target(uint8_t* data, int steer_angle_target);
 
-  // config detail: {'name': 'Steer_ANGLE_SPD', 'offset': 0.0, 'precision': 1.0,
-  // 'len': 8, 'is_signed_var': False, 'physical_range': '[0|250]', 'bit': 15,
-  // 'type': 'int', 'order': 'motorola', 'physical_unit': 'deg/s'}
+  // config detail: {'bit': 15, 'is_signed_var': False, 'len': 8, 'name':
+  // 'Steer_ANGLE_SPD', 'offset': 0.0, 'order': 'motorola', 'physical_range':
+  // '[0|250]', 'physical_unit': 'deg/s', 'precision': 1.0, 'type': 'int'}
   void set_p_steer_angle_spd(uint8_t* data, int steer_angle_spd);
 
-  // config detail: {'name': 'CheckSum_102', 'offset': 0.0, 'precision': 1.0,
-  // 'len': 8, 'is_signed_var': False, 'physical_range': '[0|255]', 'bit': 63,
-  // 'type': 'int', 'order': 'motorola', 'physical_unit': ''}
+  // config detail: {'bit': 63, 'is_signed_var': False, 'len': 8, 'name':
+  // 'CheckSum_102', 'offset': 0.0, 'order': 'motorola', 'physical_range':
+  // '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'int'}
   void set_p_checksum_102(uint8_t* data, int checksum_102);
 
  private:
   Steering_command_102::Steer_en_ctrlType steer_en_ctrl_;
-  double steer_angle_target_;
+  int steer_angle_target_;
   int steer_angle_spd_;
   int checksum_102_;
 };
